Extracted inner class VariantKeyFields outside VariantVcfFactory
